Move `Schema` and `DataType` to separate file
